fix: splashscreen
