+++ app/src/main/java/com/example/qurankarim/MainActivity.kt
@@ -1,42 +1,19 @@
 package com.example.qurankarim
 
-import android.media.Image
 import android.os.Bundle
-import android.view.Gravity
-import android.window.SplashScreen
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.width
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.geometry.Size
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.semantics.Role.Companion.Image
-import androidx.compose.ui.text.font.FontStyle
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import androidx.navigation.NavController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import com.example.qurankarim.ui.QuranListScreen
-import com.example.qurankarim.ui.theme.QuranKarimTheme
-import kotlinx.coroutines.delay
+import com.example.qurankarim.ui.SplashScreen
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -50,13 +27,12 @@ class MainActivity : ComponentActivity() {
 
 @Composable
 fun Navigation() {
-
     val navController = rememberNavController()
     NavHost(
         navController = navController,
-        startDestination = "splash_screen"
+        startDestination = Route.SplashScreen.route
     ) {
-        composable("splash_screen") {
+        composable(Route.SplashScreen.route) {
             SplashScreen(navController = navController)
         }
 
@@ -69,58 +45,9 @@ fun Navigation() {
 
             }
         }
-        composable("quran_list_screen") {
+        composable(Route.QuranListScreen.route) {
             QuranListScreen()
         }
     }
-}
-
-@Composable
-fun SplashScreen(navController: NavController) {
-
-    LaunchedEffect(key1 = true) {
-        delay(3000L)
-        navController.navigate("quran_list_screen")
-    }
-
-    Column(modifier = Modifier.fillMaxSize()) {
-        Text(
-            text = "My Qur'an",
-            color = Color.Red,
-            fontSize = 32.sp,
-            fontStyle = FontStyle.Italic
-        )
-    }
-
-    Box(
-        modifier = Modifier.fillMaxSize(),
-        contentAlignment = Alignment.Center
-    ) {
-        Image(
-            painter = painterResource(id = R.drawable.ic_background),
-            contentDescription = "",
-            modifier = Modifier.width(500.dp)
-        )
-        Box(
-            modifier = Modifier
-                .align(Alignment.Center)
-                .fillMaxWidth()
-                .width(
-                    50.dp
-                ),
-
-            )
-        Image(
-            painter = painterResource(id = R.drawable.ic_quran_logo),
-            contentDescription = "",
-            modifier = Modifier.width(200.dp)
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun SplashPreview() {
-    SplashScreen(rememberNavController())
 }
