+++ app/src/main/java/com/example/qurankarim/Route.kt
@@ -1,0 +1,6 @@
+package com.example.qurankarim
+
+sealed class Route(val route: String) {
+    object SplashScreen: Route(route = "splashscreen")
+    object QuranListScreen: Route(route = "quran_list_screen")
+}
